+++ phon/include/reset_waveform_settings_phon.hpp
@@ -3,7 +3,7 @@
 
 static const char *reset_waveform_settings_script = R"_(phon.settings.waveform = {
 	magnitude: 1.0,
-	scaling: "window"
+	scaling: "local"
 }
 )_";